Rewrite complete parser to use MLXMLNodes
This rewrites the parser to parse incoming XML into a tree of MLXMLNodes
that can be serialized back to an XML string and queried for their
information using a query language similar to prosody's find() method in
util.stanza (see https://prosody.im/doc/developers/util/stanza ).
We implement a superset of this language by extending it with some more
sophisticated features like querying the current node, extracting the
elementName of the selected node or useing ".." to ascend to the parent
of a node together with "*" and "{*}" for "any element" and "any
namespace".
Unlike prosody's find() our find: method returns an NSArray of all nodes
or extracted values matching the query string, not only the first one.
The findFirst: method can be used to get only the first node/value.
The check: method returns a BOOL that is YES if the find: method
returned one or more entries and NO if the NSArray returned by find: was
empty.
Our query language supports data conversion commands, too:
queries with extraction commands can have data conversion commands
appended, like "|bool", "|int", "|datetime" or even "|base64" which converts
the xml strings into a BOOL, NSNumber, NSDate or NSData.
The added extraction command "@@" will extract the raw attribute dictionary of
the found nodes. The extraction command "$" extracts the name of the
found element.

Co-authored-by: Friedrich Altheide <11352905+FriedrichAltheide@users.noreply.github.com>