Remove isHibernated handling in favor of a stanza queue
We reuse the outgoing smacks as general stanza queue here.
On smacks resume all stanzas queued in this queue are resent.
If a smacks resume failed or the server doesn't support smacks only
message stanzas are resent when a new session is established (e.g. after
bind) and all other stanzas (iqs and presences) are discarded.as general
stanza queue here.